fix: fix logic in ledger

+++ services/ledger/app/models/transaction.py
@@ -1,8 +1,9 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String, Numeric
 from app.database import Base
+
 
 class Transaction(Base):
     __tablename__ = "transactions"
     transaction_id = Column(Integer, primary_key=True)
     user_id = Column(String, index=True)
-    amount = Column(Float)
+    amount = Column(Numeric(10, 2))
